refactor: Refactored methods that query the identity provider into new identity service
