+++ src/Eurofurence.App.Server.Services/Abstractions/Identity/IdentityOptions.cs
@@ -1,4 +1,4 @@
-namespace Eurofurence.App.Server.Web.Identity;
+namespace Eurofurence.App.Server.Services.Abstractions.Identity;
 
 public class IdentityOptions
 {
