+++ src/Eurofurence.App.Server.Services/Users/UserService.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Eurofurence.App.Domain.Model.PushNotifications;
@@ -9,7 +8,7 @@ using Eurofurence.App.Server.Services.Abstractions.PushNotifications;
 using Eurofurence.App.Server.Services.Abstractions.Security;
 using Microsoft.EntityFrameworkCore;
 
-namespace Eurofurence.App.Server.Services.PushNotifications;
+namespace Eurofurence.App.Server.Services.Users;
 
 public class UserService : EntityServiceBase<UserRecord>, IUserService
 {
@@ -18,7 +17,6 @@ public class UserService : EntityServiceBase<UserRecord>, IUserService
     public UserService(
         AppDbContext appDbContext,
         IStorageServiceFactory storageServiceFactory,
-        AppDbContext dbContext,
         bool useSoftDelete = true
     )
         : base(appDbContext, storageServiceFactory, useSoftDelete)
