+++ src/Eurofurence.App.Server.Web/Startup/ServiceCollectionExtensions.cs
@@ -41,6 +41,8 @@ using Telegram.Bot;
 using Eurofurence.App.Server.Web.Swagger;
 using Microsoft.OpenApi.Models;
 using System;
+using Eurofurence.App.Server.Services.Abstractions.Identity;
+using Eurofurence.App.Server.Services.Identity;
 using Eurofurence.App.Server.Services.QrCode;
 using Eurofurence.App.Server.Web.Jobs;
 using Microsoft.Extensions.Logging;
@@ -84,6 +86,7 @@ namespace Eurofurence.App.Server.Web.Startup
             services.AddTransient<IDeviceIdentityService, DeviceIdentityService>();
             services.AddTransient<IUserService, UserService>();
             services.AddTransient<IArtistAlleyUserPenaltyService, ArtistAlleyUserPenaltyService>();
+            services.AddTransient<IIdentityService, IdentityService>();
             services.AddSingleton<ITelegramMessageBroker, TelegramMessageBroker>();
             services.AddSingleton<IPrivateMessageQueueService, PrivateMessageQueueService>();
